day 2

+++ run.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
-echo src/day_$1.cpp
-rm *.o
-rm out
+echo src/day_$1.cpp;
+
+if [[ -f "out" ]]; then
+    rm out
+    rm *.o
+fi
+
 g++ -c src/file.cpp -std=c++20
 g++ -c src/day_$1.cpp -std=c++20
 g++ file.o day_$1.o -o out
